requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
PyJWT==2.4.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0